+++ src/components/credential-card/credential-card.stories.ts
@@ -7,9 +7,12 @@ const meta = {
     `<d-credential-card 
     name="${args.name}" 
     issuer="${args.issuer}" 
-    description="${args.description}" 
     expiration-date="${args.expirationDate}" 
+    issued-by-label="${args.issuedByLabel}"
+    expiration-label="${args.expirationLabel}"
     verified="${args.verified}">
+    <d-badge>address</d-badge>
+    <d-badge>email</d-badge>
     </d-credential-card>`,
 } satisfies Meta<DCredentialCard>;
 
@@ -19,8 +22,9 @@ type Story = StoryObj<DCredentialCard>;
 export const Default: Story = {
   args: {
     name: 'Over 18',
-    description: 'This credential proves that you are over 18 years old',
     issuer: 'Italian Government',
+    issuedByLabel: 'issued by',
+    expirationLabel: 'exp',
     expirationDate: '2017-01-01',
     verified: false
   },